Renamed files and variables by replacing the word 'Rules' with 'Requests' for clarification purposes
