+++ src/components/Firestore/Requests/rules_evaluations_listener.ts
@@ -32,7 +32,7 @@ export function registerForRulesEvents(callback: OnEvaluationFn): Unsubscribe {
   const ws = new WebSocket('ws://localhost:8888/rules/ws');
   ws.onmessage = evt => {
     const newEvaluation: FirestoreRulesEvaluation = JSON.parse(evt.data);
-    callback({ ...newEvaluation, evaluationId: generateId() });
+    callback({ ...newEvaluation, requestId: generateId() });
   };
 
   return () => ws.close();
