+++ src/store/firestoreRequestEvaluations/actions.ts
@@ -16,7 +16,7 @@
 
 import { createAction } from 'typesafe-actions';
 
-import { FirestoreRulesEvaluation } from '../../components/Firestore/Rules/rules_evaluation_result_model';
+import { FirestoreRulesEvaluation } from '../../components/Firestore/Requests/rules_evaluation_result_model';
 
 export const addRequestEvaluation = createAction(
   '@firestoreRules/addRequestEvaluation'
