+++ src/store/index.ts
@@ -23,7 +23,10 @@ import { authReducer } from './auth/reducer';
 import { authSaga } from './auth/sagas';
 import { ConfigState, configReducer, configSaga } from './config';
 import { DatabaseState, databaseReducer, databaseSaga } from './database';
-import { FirestoreRulesState, firestoreRulesReducer } from './firestoreRules';
+import {
+  FirestoreRequestEvaluationsState,
+  firestoreRequestEvaluationsReducer,
+} from './firestoreRequestEvaluations';
 import { LogState, logReducer } from './logviewer';
 
 export interface AppState {
@@ -31,7 +34,7 @@ export interface AppState {
   database: DatabaseState;
   auth: AuthState;
   log: LogState;
-  firebaseRules: FirestoreRulesState;
+  firestoreRequestEvaluations: FirestoreRequestEvaluationsState;
 }
 
 export function* rootSaga() {
@@ -43,7 +46,7 @@ export const rootReducer = combineReducers<AppState>({
   database: databaseReducer,
   auth: authReducer,
   log: logReducer,
-  firebaseRules: firestoreRulesReducer,
+  firestoreRequestEvaluations: firestoreRequestEvaluationsReducer,
 });
 
 export function createStructuredSelector<T>(
